chore: rename wallpaper.domain
