+++ frontend/core/stores/wallpaper/hooks.tsx
